Configurable support for 64-bit ACPI integers (ACPI 2.0)
date	2000.10.11.21.02.00;	author rmoore1;	state Exp;

+++ source/components/executer/excreate.c
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
  * Module Name: amcreate - Named object creation
- *              $Revision: 1.47 $
+ *              $Revision: 1.48 $
  *
  *****************************************************************************/
 
@@ -365,7 +365,7 @@ AcpiAmlExecCreateField_original (
     }
 
 
-    Offset = OffDesc->Number.Value;
+    Offset = (UINT32) OffDesc->Number.Value;
 
 
     /*
